+++ src/main/java/com/js/calendar/repository/DayRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DayRepository extends PagingAndSortingRepository<Day, Integer> {
+public interface DayRepository extends PagingAndSortingRepository<Day, Long> {
 
 }
